+++ pkg/webkit/webview_cgo.go
@@ -2580,6 +2580,7 @@ func (w *WebView) ID() string {
 	return strconv.FormatUint(uint64(w.id), 10)
 }
 
+
 // PrefersDarkTheme returns true if GTK is configured to prefer dark theme
 func PrefersDarkTheme() bool {
 	return C.gtk_prefers_dark() != 0